refactor(chiptune): rework amplitude normalization around a 32-bit mixed wave output
- replace chiptune_fetch_64bit_wave() with chiptune_fetch_32bit_wave()
  so the mixed wave is reduced to int32_t before the final normalization step
- rename the internal amplitude normalization state from gain to divisor
  to match the actual / divisor implementation; this is a naming cleanup and
  does not change the divisor-based calculation logic itself
- update the scaled reciprocal multiplier path to fit the new int32_t mixed
  wave output, including changing the reciprocal multiplier and related
  calculations to int32_t where appropriate
- keep the public API and GUI in gain terms, while mapping
  (UINT8_MAX + 1) - gain to the internal normalization divisor
- simplify the amplitude gain slider to use the direct 1..255 gain range
  instead of the old reversed 16-bit-style range

+++ chiptune.h
@@ -71,8 +71,8 @@ int chiptune_set_melodic_channel_timbre(int8_t const channel_index, int8_t const
 									  float const envelope_damper_sustain_duration_in_seconds);
 
 
-int32_t chiptune_get_amplitude_gain(void);
-void chiptune_set_amplitude_gain(int32_t const amplitude_gain);
+uint8_t chiptune_get_amplitude_gain(void);
+void chiptune_set_amplitude_gain(uint8_t const amplitude_gain);
 #ifdef __cplusplus
 }
 #endif
